Lock bundler version to 2.1.4
Travis builds started failing on 10 Dec 2020.

The failure occurs when installing the gem `libv8`: `Failed to build gem native extension.`

This coincided with bundler version 2.2.0 becoming the default, instead of 2.1.4.

Our Dockerfile didn’t specify a version.

I propose to resolve this in the first instance by locking bundler to 2.1.4, known to work with libv8.