Remove Activity#planned_disbursements
This association is no longer used, and _should_ not be used. Views
wanting to show the forecasts for an activity should use the
`latest_planned_disbursements` method, which uses the
PlannedDisbursementOverview service to return only the latest values for
each quarter, rather than _all_ the historical records linked to the
activity.

Removing this association should discourage people from using the wrong
method for displaying forecasts.