Importer handles no implementing organisation
I'll be honest and say I am struggling to follow how the
ImplementingOrganisationBuilder is working, but I do know that when
trying to import real data to production we have some issues.

I've written the tests to reproduce the issues and get around them,
although I feel like there is probably a better approach than this, with
time in short supply to get these imports done I think this potential
work-around is good enough.

I believe the issues are:

When the importer is creating an activity it assumes the import row will
have the details to make an implementing organisation and this is not
always the case.

When the importer is updating an existing activity it assumes the
activity has an existing implementing organisation or that the import
row has the details to create one, again this is not always the case.

When updating an activity we only add the result of trying to build an
implementing organisation if it is valid, for an implementing
organisation to be valid it will need the activity_id, name and type.

We cannot use this approach for creating new activities because the
activity will not have an id until it is saved.

It looks like we create a new activity and implementing organisation at the
same time and the implementing organisation will only become valid once
the new activity is saved and able to give the implementing organisation
its `activity_id`. Because these two things happen at the point the
activity is saved, it may be we are trying to save an invalid
implementing organisation, there is no way to know until the entire
activity is saved.

The fix here is if  Implementing organisation name or type is present
in the row we have to assume the user is trying to set an implementing
organisation for this row, and if one of the values is missing we need to tell them
with an error.

But if neither is present, we don't want to try and create an implementing
organisation at all.

We also put a guard in to stop trying to add errors to the
ImplementingOrganisationBuilder if there are no implementing
organisations.