Include the next twelve quarters' forecasts in exports
`Report#next_four_financial_quarters`, and the methods based on it in
ExportActivityToCsv, return the next four financial quarters and we
generate forecast values for those in the exported CSV.

We want to expand how far ahead we present forecasts to by changing from
four to twelve quarters. This commit does that and changes the names of
all related methods.