Return future quarters from the report's date, not today's date
`Report#next_twelve_financial_quarters` returns the next twelve quarters
from today's date, not from the date of the report. That means the
result of ExportActivityToCsv depends on when it's executed, not on the
date of the requested report.

We want the exported data to always include future forecasts starting
from the date of the requested report, regardless of the current date.

The tests reflect this by restricting the `travel_to` calls to wrap the
creation of the Report object, not the execution of the entire test, so
it affects the date on the report but does _not_ change what the rest of
the test believes the current date to be.