CSV template now shows activity title
On the CSV template for bulk uploads, we were showing the activity description
for the column `activity name`. The activity title would be more appropiate in
this case.
The activity title is only shown here to help guide users, we do not use this
column from the template during import.

Specs have been modified accordingly.