Create validator for `planned_end_date`
Until now, when a user add a `planned_end_date` that is earlier than the
`planned_start_date`, RODA allows this to happen, which makes more likely for
users to accidentally report incorrect information. We need to validate that
the planned end date will always be in the future (or at least on the same day)
than the start date. We also need to warn the users with an error message if
they introduced incorrect dates.