Run Docker-compose in Github Actions
As well as adding the Github actions config, this also adds a simple `script/cibuild`
script to run the test script in Docker. I've had to make a couple of tweaks to get
our Docker config working properly in CI, namely removing the volumes from the
docker-compose.test.yml, and making sure `shellcheck` is intalled in the test layer.