Cache the Docker container
This uses Github actions' cache functionality to attempt to find a
cached version of the Docker container. If it does, it loads the cached
version, to cut down on build times. If there is no cached version
available (i.e. the Dockerfile has changed), it uploads the latest version
at the end of the test run