Helper method to validate date inputs
We need to check if the date parts that the user inputs can form a valid date.

It’s not an error if all the parts are blank, but if any parts are either missing, or do not constitute a valid date, we want to raise a specific error.

We are going to use this method in a future commit.