Show an error message if user inputs invalid date
We want to be able to recognise when the user has tried to input an invalid date, and communicate that to them.

When Rails processes multipart date parameters, it ignores any input where one or more of the parts are missing. This means that the invalid input is silently ignored.

When all the parts are present, but they do not form a valid calendar date, ActiveRecord raises a `MultiparameterAssignmentErrors` exception.

Previously, the application dealt with this by processing the date parts and silently ignoring any invalid dates.

However, this creates confusion and uncertainty for users, who would prefer to be alerted that they tried to enter an invalid date.

Instead of ignoring, we are now detecting any malformed date inputs, and stopping the processing of the form to show the error.

There are a few things that might require explanation:

1. Why are we not using a custom model validation?

Model validations rely on the attributes being assigned to the object, so they can be verified. However, ActiveModel rejects the assignment when the resulting date would be invalid, so none of the date parts make it to the model.

2. Calling `valid?` on the activity object wipes any existing errors, so once we’ve added our custom errors to any date fields that need them, we need to render the form step without attempting to save the activity.

This is why we use `render_step` instead of `render_wizard @activity`.