`oda_eligibility` vendor file amended
For consistency throughout the service, we are now going to codify the options
for `oda_eligibility`, using numbers instead of strings.

I have modified the yaml file to align with this new codification. I have also
placed the file in the right vendor folder, since these values are BEIS specific,
not IATI.