Adapt import work for new `oda_eligibility` codes
We want to be consistent on the service and ask users to provide codes when we
can, instead of strings. Until now, `oda_eligibility` on the CSV import work was
taken as a string. We are now collecting the numeric values (since this field is
an enum on the DB). This commit reflects those changes and amends the specs so
we have the certainty that we get the desired behaviour from the importer.