Show error when cell contains invalid characters
At the moment, if a cell contains an invalid (i.e. non-UTF-8) character, we get a
500 error, because the template code attempts to show the value of the cell, and
this raises an `Encoding::CompatibilityError`. We can get around this by catching
that specific error and forcing the encoding of the cell. We can also display a
slightly more friendly error back to the user.

# Conflicts:
#	CHANGELOG.md