Cache against dependencies
This adds the Gemfile.lock and package-lock.json files to calculate the validity of
the cache. At the moment, the cache is only refreshed if the Dockerfile changes,
meaning that if a Ruby / JS dependency changes, we have to run the whole bundle
install again when building the container. This shaves a good 3 - 4 minutes off
subsequent builds.