Ensure we get an error for two-digit years
Since `strptime(date, "%d/%m/%Y")` will parse two-digits years as-is, we
need to make sure these generate an error.

Luckily, this is already trapped by the `between` validation on `date`
(which must currently be between 10 years in the past and 25 years in
the future), but I felt like it still warranted a test case.