Bug fix for `N/A` option in SDGs
Up until now, when a user selected that the SDGs were not applicable to a
certain activity, the activity details page would now reflect this. Instead, the
section for SDGs would appear blank and we could still see the edit button as
`Add`. This commit introduces changes on the activity presenter, so when
sdgs_apply is set to false, the user will see `Not applicable` on the activity
details page, and the edit button will appear as `Edit`.