Include current financial year in “next ten years”
The intention of the method was always to include the current year in the “next ten years” list.

However, what was missing was the current _financial_ year, rather than the calendar year.

Fix to always include the current financial year in the list, which is used when creating a forecasted spend via web form.