Activity calculates the IATI status from programme status
The `status` field, used to store the IATI status code, is always calculated from the `programme_status` field, and never set explicitly via the interface. It is only needed when exporting data for IATI publishing.

As it is a derived value, we don't need to persist it to the database, so we can simplify our schema.

For clarity, the new method is named `iati_status` rather than `status`.