Remove form steps label from child activities view
This is an old feature that shows if a user has completed all the form steps of
an activity or not, which used to be an indicator of an activity being valid.
This is no longer the case and this information has no value for the users
anymore, so we are now removing this column from the child activities view.
This applies to programmes, projects and third-party projects.

I have also removed the specs related to this feature.