Importer ignores incomplete activities when assigning a parent
Raise an `invalid_parent` error if the parent being associated with a
new activity is incomplete (`form_state` isn't `completed`).

On production, we have several activities in an incomplete state, which
meant that `Activity.by_roda_identifier` would return a result, instead
of `nil`, so `parent_not_found` was not raised.

This would lead to these activities being imported and incorrectly
parented to a random activity.

Finally, the test case behaved which made writing this more complicated
than it should've been. Because `let` blocks are lazily-evaluated,
calling `new_activity_attributes` triggers the creation of
`existing_activity` which led to the `to change { ... }` expectation
always being off by one.

This was resolved by changing to a `let!` for `existing_activity`. I've
added a comment to prevent someone else making that mistake in the
future!