Fix spec that would fail for the wrong reason
The spec is exercising the ordering of the budgets, so the period of time that the budgets are for isn’t relevant.

The spec would fail when there was a leap year involved, because the validator intentionally checks for 365 days, not the variable duration of “a year”. (This was a choice, well documented in the commit message 6a9d56e44.)

Remove the brittle spec arrangement and replace with a time period that would always be less than 365 days.