Fix: Users can change the extending organisation
Whilst looking at if this was even possible I noticed it was broken.

The reason is that the activity model is not valid by default without
all required attributes. We often see activities being completed over
time and so cannot always rely on the object being valid.

It turns out we even have a validation context that was used for these
reasons to test the validity but it was no called on the `save` so the
save never completed.

This was compounded by the fact the UI gives the impression that the
save was completed when it had not.

The fix is to simply validation in the `update_extending_organisation`
context which will only validate the extending organisation attribute
and not the entire Activity object.