Infer if FSTC applies from aid type where possible
For some values of aid type, it is determined automatically whether FSTC can apply or not. In those cases, calculate it and skip the step.

In the case of aid type C01, we want to help the user by pre-selecting the most common option, but still allow them to specify if that’s not the case.

Nothing changes for all the other supported aid types.