Don’t show edit link if FSTC inferred from aid type
If the value of FSTC applies is inferred from the value of aid type, the user has to edit the aid type before they can change the FSTC applies answer.

Trying to edit FSTC directly didn’t work, as we skip the step, so I chose to remove the edit link for those situations.