Refactor `activity_forms_controller`
While working on a chore in relation with `form_state` I decided to move the
FORM_STEPS constant to the Activity model. This was previously in the
activity_forms_controller but it makes more sense to have it in the model, so
the controller will communicate with the model rather than the other way around.

This commit also contains route modifications on helpers and specs now that the
FORM_STEPS live somewhere else.