Prevent `form_state` from being set to a wrong value
We recently had an issue where after a data migration we accidentally set the
value of `form_state` to `completed` instead of `complete` for some activities
on production. There was no error raised because this field lacked of any
validation. To avoid similar errors in the future we are now introducing
validation to only accept one of the expected values. These values are all the
FORM_STEPS plus some custom ones. We also allow nil.