Allow Transactions to be deleted
This adds a controller action to delete a transaction. It also creates a PublicActivity
activity. This happens before the transaction is actually deleted, because we can't
call `transaction.create_activity` after the action has been deleted. I don't think
this matters too much because we've carried out all the checks that the
transaction can be deleted by this point, so if the transaction can't be deleted at
this point, we've got bigger problems.