Add a `.clear!` method to delete a forecast
After a bit of back and forth and discussion, we decided that when a user wants
to "delete" a forecast, it'll be because they've erroneously created one, either for
the wrong reporting period, or the wrong activity, rather than they've just entered
the wrong amount. With this in mind, I decided that the expected behaviour would
be to delete all the Planned Disbursements for that activity / period, and remove
all history. We did discuss just setting the latest value to zero, which would behave
as if the forecast was deleted, but this would still keep the erroneously created
value as the "original", which may not be the intent.