Refactor ActivityFormsController
The ActivityFormsController has a lot of repetition and makes the controller quite
large. I've factored all the update actions into a lean(er) class that accepts the
activity and the specific step, then works out which parameters to fetch and apply,
as well as any extra steps that need to be performed. I haven't bothered with any
unit tests for this class, which does make me a little anxious, but I think it's fine, as
this is all covered by feature tests.