Users can upload activity data from CSV
Opens up the existing activity CSV importer to users.

The importer expects a CSV file with human-friendly headers but coded attribute values (for example, the Activity Status must be "7" instead of "Spend in Progress").

Errors are surfaced in the view in the same way that we surface them when users upload actuals.

We are offering a template file, which only contains the headers.

We are not trying to detect which fund this upload applies to, so all the columns are present, but they will only be required as applicable for each project.