Separate organisations in `users_controller`
Until now the `users_controller` would pass an instance variable with all the
existing organisations on the database, based on the policy scope the user was
under. Now, as part of a piece of work we are going to implement in the next
commit, I have divided these organisations into service_owner (BEIS) and
Delivery Partners.