Separate BEIS from other orgs on user admin page
When a user wants to add a new user, they have to select what organisation this
user belongs to. There is a radio button list for this, with organisations
sorted alphabetically. However, there was no distinction between BEIS and the
Delivery Partners on this list. BEIS appeared in the middle of the list, which
made easier for users to accidentally assign a new user to BEIS when they don't
belong to this organisation. This would give that user higher privileges than
they should have. To avoid this, we have now separated BEIS org from the rest of
organisations on display, adding a divider in the radio button list.