Parent step only shows complete parent activities
Previously when prompting for a parent activity, all parent activities were shown
on the form. This resulted in a bunch of incomplete or erroneously created
activities showing up. This adds a filter to the `scoped_parent_activities` method
that makes sure only activities with a `form_state` of `complete` are shown.