Remove the ability to set the Report used by ImportPlannedDisbursements
In normal operation, PlannedDisbursementHistory adds forecast data to
the current editable Report for the given Activity. This service allowed
a different Report to be passed in by the caller, for the purposes of
adding data to an older closed report rather than the current open one.

This was for a one-off task of backfilling data into an old report and
is no longer needed, and should definitely not be exposed to end users.
Therefore we can remove this parameter and any functionality attached to
it.