Provide a user-facing interface for bulk uploading forecasts
This exposes the functionality of ImportPlannedDisbursements to end
users via the website. It is closely modelled on the same workflow for
transactions.

The user begins by downloading a spreadsheet template containing a list
of all the activities covered by the current report. This template lists
the name, delivery partner ID, and RODA ID for each activity.

After these three columns, the user must add more columns for thw
quarters they wish to add forecasts for. The format for the column name
is `FC <year> FY Q<quarter>`, for example:

    FC 2020/21 FY Q3

This is based on the format of the spreadsheets we were given to
backfill the forecasts for 2020/21 Q2. There are some usability
questions around this design and I expect we'll revise it; this commit
is purely designed to expose the existing functionality on the website.