Use Error objects in ImportPlannedDisbursements
This importer was originally designed to be invoked by a command-line
program and so its errors were just strings to be printed to the
terminal.

To make this class usable from the web, we need to expose error
information in a structured way so we can display it in a table. This
borrows the Error model from ImportTransactions to represent the cell in
a CSV file that caused the error.