Refactor template generation for planned disbursement uploads
We would like to add some validation logic to make sure that users only
add column names that match the format for identifying the financial
year and quarter, and any other columns are treated as errors.

T odo this, ImportPlannedDisbursements needs to know all the valid
column names that are in the template to begin with, so we move those
names out of Staff::PlannedDisbursementUploadsController and into
ImportPlannedDisbursements::Generator. Since the latter knows what the
column headings are, it should also be responsible for generating the
data tuple for each activity.