Raise an error when a forecast upload contains unrecognised columns
The bulk upload feature for forecasts relies on the user adding columns
to the spreadsheet for each quarter they want to report on. Those
columns must have the format:

    FC <year> FY Q<quarter>

If the user adds a column that does not match this format, we should
raise an error rather than letting the upload appear to succeed when we
may have ignored some of its contents.

As the error relates to a column heading, not an individual cell, we
maintain a set of column headings we've already raised an error about,
so that we don't emit this error for every row in the upload. The test
contains two identical rows to make sure only one error is reported
about the column heading.