Add a Codelist model
This adds a Codelist model which loads all codelists into memory on the first
initialization in production. It dev, all codelists are loaded on demand, which makes
it easier to make changes on the fly.

The presence of this model also increases the visibilty of codelists across the
application, making it easier for us to use them in things like controllers and
models, as they're not just scoped to a helper.