Restore progress-style RSpec output locally
Whilst the 'documentation' style formatting works better on CI, it can
be pretty distracting when running in development.

This removes the change globally, apart from when running in CI or when
running a single test (using RSpec's recommended config).  Developers
may choose to override this globally, by adding `--format=documentation`
to `~/.rspec`.