Add Transfer model
This includes reference to `source` and `destination` which are the
activities on either end of the transfer.

I've added foreign key constraints, to prevent activities from
being deleted when a transfer references them.

Finally, whilst the Trello card originally asked for an `amount` column,
I've opted to use `value` for consistency with the existing tables.