Add new route for historic activities
We are going to implement a new feature that separates the activities view page
into two different tabs. One containing the `Current` activities, which are
more relevant for users, and the other containing the `Historic` activities,
which are the ones where users are unlikely to add any financials, information,
etc to them.

This commit adds the new route for the Historic activities tab. We will be using
the current `/activities` route for the Current activities.