Controller actions for activities view page
As mentioned in the previous commit, we are separating the activities into
Current and Historic. The users focus their attention on the current ones, so
we decided to separate the ones that are not so relevant for them. These
`historic activities` are the ones in which `programme_status` is set to either
`completed`, `cancelled` or `stopped`. These activities cannot have any
financial data added or modified and the users are unlikely to have to use them
in the future. However, we still want to keep them in RODA.

The activities view will display two tabs (see next commit) `Current` and
`Historic`, and the user will be able to select which ones they want to see.

This commit introduces the changes on the controller and new scopes on the model