Fixing BEIS filter view
As a BEIS user, you can filter activities for each delivery partner. When we
introduced the two tabs view for activities, we had to make a small tweak on the
way the filter displays the information. Now, when a BEIS user filters
activities for a DP, they will see first the tab in which they are when they
decided to filter the activities. That way, they will be able to see both
Current and Historic activities.

There is test coverage for this feature in this commit.