Add a way to group projects by parent project
This method will be used in a subsequent commit to group and order projects and third-party projects in the report view.

The design for the report did not specify any sorting order, but there are designs for the activity index view that specify that “- Activities within a level are ordered by Identifier”.

We like consistency, so we chose to sort them for the report view as well.

Co-authored-by: jcoglan <james.coglan@dxw.com>