Remove period start and end dates
Now we work out the start and end of a period from the financial year, we no
longer need to ask for the start and end dates, as they can be calculated and
returned on the fly. We also no longer need to ask users for them, so we can
remove them from the UI. We will remove the fields from the database at a later
date, once we've deployed these changes and are sure the data migration has run.