Add Financial Year and Quarter classes
This allows us to get the beginning and end of a financial year, given the calendar
year that the financial year starts in, as well as the Financial Year and Quarter for
a given date. I noticed there were some bits of shared behavior in another module,
so I took the oppotunity to do a bit of a refactor.