Set DATABASE_CLEANER_ALLOW_REMOTE_DATABASE_URL during assets:precompile
In order to upgrade to database_cleaner 2.0, we need to fix an error
that happens while loading the app. We get the following exception
during the CI build when running this gem:

    DatabaseCleaner::Safeguard::Error::RemoteDatabaseUrl:
    ENV['DATABASE_URL'] is set to a remote URL. Please refer to
    https://github.com/DatabaseCleaner/database_cleaner#safeguards

That URL gives two ways to turn this safeguard off. You can set the
environment variable `DATABASE_CLEANER_ALLOW_REMOTE_DATABASE_URL=true`,
or do this in your application setup code:

    DatabaseCleaner.allow_remote_database_url = true

I think this safeguard is worth having and I would like to only turn it
off in the CI environment, where we know it's safe. I'd therefore rather
not turn this off in the `test` environment generally, as that would
allow devs running tests locally to accidentally wipe a remote DB.

As it turns out, that option is out of the question anyway, because of
where the exception comes from: it's triggered by this call in
`config/application.rb`:

    Bundler.require(*Rails.groups)

That is, it's caused by just loading all our gems. We can't configure
`DatabaseCleaner` before it's loaded, and putting something in an
initialiser or in `config/environments/test.rb` would be too late to
catch this error. Additionally, the exception is triggered while running
the `assets:precompile` task during our Docker image build, so it's not
even running a test-specific task that should need to load any backend
code.

This environment variable is already set in `docker-compose.test.yml`,
which means it's set when the container boots to run the application,
but _not_ while the image is being built. Therefore the
`assets:precompile` task doesn't have it set and the app fails to load
as a result.

To fix this, I am adding the environment variable to the
`assets:precompile` task in Dockerfile. I don't like the duplication,
but I don't know another way to set that variable; environment variables
set by docker-compose aren't visible during an image build, and neither
are thoese set in .github/workflows/ci.yml.

I also don't know where merely loading our gems causes this exception,
before `DatabaseCleaner` is loaded or called, and before there's any
opportunity to configure it. I'm just doing this to get the build
working so the upgrade can be merged.