Controller concern for sending downloadable CSV files
We have a few places in the application where we send a CSV file as a
stream with `Content-Disposition: attachment` which involves a fair bit
of boilerplate. We also have discovered that the encoding we use for
these files, UTF-8, does not open successfully in Excel -- many bytes
are misinterpreted using Windows 1252, MacOS Roman, or some other
single-byte encoding.

Therefore I thought I would extract the logic these downloads into a
mixin, so we have one place where the semantics are defined. This takes
care of setting the required HTTP headers and sending each line of
output, using a given encoding. You would use it as follows:

    stream_csv_download(filename: "report.csv", headers: [...]) do |csv|
      report_data.each do |row|
        csv << row
      end
    end

The `headers` argument is distinct from the data rows so that we have
the option of prepending a byte order mark (BOM, the codepoint U+FEFF)
to the first line of output.

We have run some experiments using various encodings of the same data to
try to determine which encoding to use by default. We have tried UTF-8,
UTF-16BE and UTF-16LE, all with and without a BOM at the beginning of
the file. On Windows, it appears that Excel will open UTF-8 with a BOM.

If given a UTF-8 file with no BOM, Excel will completely mis-read the
file, interpreting the wrong codepoints for most Unicode characters.
This is despite the fact that the BOM character is intended for telling
the difference between UTF-16BE and UTF-16LE and should have no meaning
in UTF-8.

Given a UTF-16LE file with a BOM, we found that Excel would interpret
most characters correctly but would ignore commas, so it thought the
file contained a single column.

Therefore I am using UTF-8 here, with a BOM, just because I think adding
BOM makes the file less ambiguous for most purposes.