Use StreamCsvDownload for sending bulk upload templates
This replaces the boilerplate for sending bulk upload CSV templates of
various types with a call to `stream_csv_download`. This encodes the
output as UTF-8 and prepends a byte order mark.

Ruby's CSV parser treats these BOM bytes as data, and does not use them
to infer the data's encoding, so during tests they must be removed
before parsing, and the data must be explicitly transcoded to UTF-8.
Otherwise the BOM ends up being treated as part of a column header.