Use StreamCsvDownload for sending reports
Here we replace the boilerplate for sending downloadable report CSV with
a call to `stream_csv_download`. This means the resulting data is
encoded as UTF-8, consistent with the other CSV downloads we offer.

Since StreamCsvDownload deals with text encoding, the
ExportActivityToCsv service should not return arrays of data and not
pre-serialise them to CSV.