Revert our Database Cleaner fix
When DatabaseCleaner 2.0 was released, we had to nail a fix into our Dockerfile,
as DatabaseCleaner was alerting us that we were using a remote URL for our
database at the point of the app loading. We already have configuration to allow
this to happen in our test suite, but this does not get loaded in time, so we had to
add a env var to our Dockerfile to fix this.

The release of 2.0.1 has updated the code to load the safeguards later in the
initialization process, so I think it's worth getting rid of this fix now.