Backing services can be run with Docker
I prefer to run the two backing services (Postgres and Redis) with
Docker as I find managing version with Homebrew can effect my entire
system and I often find I update the backing services without meaning to
when updating my own tooling, which can cause horrible problems.

For those that prefer it, a simple docker compose file allows one to
start and stop the backing services via docker and not effect their
system at all. This has an added benefit of having the option to pin the
versions of the backing services to that of the ones used in production.

Whilst I am loathe to add alternatives to the one-script-to-rule-them-all,
I much prefer this approach personally and have real technical
reasons to avoid the Homebrew path, therefore I am adding this compose
file and a small section in the readme to cover its use for those that
prefer.