Add CreateProgramme service
Create a new programme-level activity associated with the given
organisation and source fund.

This will eventually replace the combination of `CreateActivity` and
`UpdateActivityAsProgramme` currently in use.

Currently, this sets `extending_organisation` to BEIS, which mirrors the
existing behaviour. This is wrong, but will be fixed across the app in a
separate PR.

NB: In order that the test can check that the `source_fund` has been set
to the fund activity's `source_fund`, I've modified the Fund class'
equality (`==`) method. Without this, two `source_fund`s would not be
equal because they each map to a different instance of `Fund.new(x)`.