Introduce ActivityDefaults service object
This accepts a `parent_activity` and `delivery_partner_organisation` and
returns a hash of default attributes ready to be applied to a new
`Activity`.

To start with, this will be used when creating new child activities, but
this is probably useful to DRY up other parts of the app (such as during
activity bulk-upload).